+++ packages/bluesky/lib/src/services/services.dart
@@ -34,38 +34,49 @@ sealed class BlueskyService {
         mockedPostClient: mockedPostClient,
       );
 
-  /// Returns the actors service.
+  /// Returns the actor service.
+  /// This service represents `app.bsky.actor.*`.
   ActorsService get actors;
 
-  /// Returns the feeds service.
+  /// Returns the feed service.
+  /// This service represents `app.bsky.feed.*`.
   FeedsService get feeds;
 
-  /// Returns the notifications service.
+  /// Returns the notification service.
+  /// This service represents `app.bsky.notification.*`.
   NotificationsService get notifications;
 
-  /// Returns the graphs service.
+  /// Returns the graph service.
+  /// This service represents `app.bsky.graph.*`.
   GraphsService get graphs;
 
   /// Returns the unspecced service.
+  /// This service represents `app.bsky.unspecced.*`.
   UnspeccedService get unspecced;
 
-  /// Returns the servers service.
-  atp.ServersService get servers;
+  /// Returns the server service.
+  /// This service represents `com.atproto.server.*`.
+  atp.ServerService get servers;
 
-  /// Returns the identities service.
-  atp.IdentitiesService get identities;
+  /// Returns the identity service.
+  /// This service represents `com.atproto.identity.*`.
+  atp.IdentityService get identities;
 
-  /// Returns the repositories service.
-  atp.RepositoriesService get repositories;
+  /// Returns the repo service.
+  /// This service represents `com.atproto.repo.*`.
+  atp.RepoService get repositories;
 
   /// Returns the moderation service.
+  /// This service represents `com.atproto.moderation.*`.
   atp.ModerationService get moderation;
 
   /// Returns the sync service.
+  /// This service represents `com.atproto.sync.*`.
   atp.SyncService get sync;
 
-  /// Returns the labels service.
-  atp.LabelsService get labels;
+  /// Returns the label service.
+  /// This service represents `com.atproto.label.*`.
+  atp.LabelService get labels;
 }
 
 final class _BlueskyService implements BlueskyService {
@@ -146,13 +157,13 @@ final class _BlueskyService implements BlueskyService {
   final UnspeccedService unspecced;
 
   @override
-  final atp.ServersService servers;
+  final atp.ServerService servers;
 
   @override
-  final atp.IdentitiesService identities;
+  final atp.IdentityService identities;
 
   @override
-  final atp.RepositoriesService repositories;
+  final atp.RepoService repositories;
 
   @override
   final atp.ModerationService moderation;
@@ -161,5 +172,5 @@ final class _BlueskyService implements BlueskyService {
   final atp.SyncService sync;
 
   @override
-  final atp.LabelsService labels;
+  final atp.LabelService labels;
 }